+++ chrome-extension/src/features/utils/octokitRepoUrl.ts
@@ -1,0 +1,5 @@
+import { createName } from "./createName";
+
+export const octokitRepoUrl = (repo: string) =>
+  // get repo name for octokit
+  `/repos/${createName(repo)}`;
