+++ chrome-extension/src/features/utils/popupCenter.ts
@@ -1,0 +1,48 @@
+export const popupCenter = ({
+  url,
+  title,
+  w,
+  h,
+}: {
+  url: string;
+  title: string;
+  w: number;
+  h: number;
+}) => {
+  // Fixes dual-screen position                             Most browsers      Firefox
+  const dualScreenLeft =
+    window.screenLeft !== undefined ? window.screenLeft : window.screenX;
+  const dualScreenTop =
+    window.screenTop !== undefined ? window.screenTop : window.screenY;
+
+  const width =
+    window.innerWidth ||
+    document.documentElement.clientWidth ||
+    window.screen.width;
+  const height =
+    window.innerHeight ||
+    document.documentElement.clientHeight ||
+    window.screen.height;
+
+  const systemZoom = width / window.screen.availWidth;
+  const left = (width - w) / 2 / systemZoom + dualScreenLeft;
+  const top = (height - h) / 2 / systemZoom + dualScreenTop;
+
+  const newWindow = window.open(
+    url,
+    title,
+    `
+      scrollbars=yes,
+      width=${w / systemZoom},
+      height=${h / systemZoom},
+      top=${top},
+      left=${left}
+      `
+  );
+
+  if (newWindow && newWindow !== null) {
+    newWindow.focus();
+  }
+
+  return newWindow;
+};
