+++ chrome-extension/src/features/utils/timeout.ts
@@ -1,0 +1,4 @@
+export const timeout = (ms: number) =>
+  new Promise((resolve) => {
+    setTimeout(resolve, ms);
+  });
